Feat(Electron): Create a Electron Bundled Project
Create an Electron, React & Typescript application that is accessible for mac, windows and linux. App is currently functional on Mac as dmgs have been downloaded from the app which lie within the dist folder. Shell command utilised to create the mc bundle is `npm run dist:mac` Other commands for other OS can be found in the package.json (dist:win & dist:linux). gitignore applied to dist-electron folder in order to protect the sensitive information of the manner in which the electron app has been structured within main.js.