Refactor(File Extension): Change .jsx to .tsx
Error occured when changing file to .tsx specifically in the handleTabChange Event , Typescript complained about. 'e' is declared but its value is never read.ts(6133) and
Parameter 'val' implicitly has an 'any' type.ts(7006). Typescript does this in order to maintain type safe code. Fix is to declare the data type of 'e' and 'val'