Refactor(File Extension): Change .jsx to .tsx
Change file extension from .jsx to .tsx in order to eliminate the typescript import errors within the app.tsx file, a .tsx file is a typescript file that supports .jsx syntax.

Error occur when changing file to .tsx specifically in the handleTabChange Event , Typescript complained about. 'e' is declared but its value is never read.ts(6133) and
Parameter 'val' implicitly has an 'any' type.ts(7006). Typescript does this in order to maintain type safe code. Fix is to declare the data type of 'e' and 'val'
6ec1347