Creation of Ingredients panel and editing of AddStock and Deliveries
In this commit i have included the code for the new file Ingredients. in which is the ingredients tab of the table belonging to the Productions page. I have edited the AddStock page to use the newly created component of the warehouseto button that is a functional drop down button.
In the Deliveries page i have simply adjusted the padding of the table contents